Functions moved to turnController
